connected api to frontend/client/comment add
