+++ client/index.html
@@ -19,7 +19,7 @@
               <ul class="navbar-nav ms-auto text-uppercase">
                   <li class="nav-item"><a class="nav-link" href="#about">about me</a></li>
                   <li class="nav-item"><a class="nav-link" href="#portfolio">competency</a></li>
-                  <li class="nav-item"><a class="nav-link" href="#review">reviews</a></li>
+                  <li class="nav-item"><a class="nav-link" href="#review">comments</a></li>
                   <li class="nav-item"></li>
                   <li class="nav-item"><a class="nav-link" href="#contact">Contact</a></li>
               </ul>
@@ -31,7 +31,7 @@
 
     <script type="module" src="/src/main.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0/dist/js/bootstrap.bundle.min.js"></script>
-    <script src="src/assets/assets/agency.js"></script>>
+    <script src="src/assets/assets/js/agency.js"></script>
     <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
     <script src="https://cdn.jsdelivr.net/npm/popper.js@1.14.3/dist/umd/popper.min.js" integrity="sha384-ZMP7rVo3mIykV+2+9J3UJ46jBk0WLaUAdn689aCwoqbBJiSnjAK/l8WvCWPIPm49" crossorigin="anonymous"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@4.1.3/dist/js/bootstrap.min.js" integrity="sha384-ChfqqxuZUCnJSK3+MXmPNIyE6ZbWh2IMqE241rYiqJxyMiZ6OW/JmZQ5stwEULTy" crossorigin="anonymous"></script>
