+++ client/src/assets/base.css
@@ -5348,6 +5348,7 @@ textarea.form-control-lg{
     display:inline-block;
     width:2rem;
     height:2rem;
+
     background-repeat:no-repeat;
     background-position:50%;
     background-size:100% 100%
